Fix map update, replace existing maps on update (#9598)
Map update was written to not expect an unzipped map to already
exist. In the case of map update, then there will be an already
existing map folder. This update fixes the update to do a replace
of any existing content.

The folder replace is done with a backup and rollback if it fails.
Before we replace the old map, we will move it to temp. If the
move of the new map fails, then we will move the old map
from the temp backup to its original location.

+++ game-app/game-core/src/main/java/games/strategy/engine/framework/map/file/system/loader/ZippedMapsExtractor.java
@@ -95,6 +95,11 @@ public class ZippedMapsExtractor {
    * @return Returns extracted location (if successful, otherwise empty)
    */
   public static Optional<Path> unzipMap(final Path mapZip) {
+    Preconditions.checkArgument(!Files.isDirectory(mapZip), mapZip.toAbsolutePath());
+    Preconditions.checkArgument(Files.exists(mapZip), mapZip.toAbsolutePath());
+    Preconditions.checkArgument(
+        mapZip.getFileName().toString().endsWith(".zip"), mapZip.toAbsolutePath());
+
     try {
       return unzipMapThrowing(mapZip);
     } catch (final IOException e) {
@@ -103,21 +108,14 @@ public class ZippedMapsExtractor {
     }
   }
 
+  // unzip map
+  // if previous folder exists then back it up
   private static Optional<Path> unzipMapThrowing(final Path mapZip) throws IOException {
-    Preconditions.checkArgument(!Files.isDirectory(mapZip), mapZip.toAbsolutePath());
-    Preconditions.checkArgument(Files.exists(mapZip), mapZip.toAbsolutePath());
-    Preconditions.checkArgument(
-        mapZip.getFileName().toString().endsWith(".zip"), mapZip.toAbsolutePath());
-
-    final String extractionFolderName = createExtractionFolderName(mapZip.getFileName().toString());
+    // extraction target is important, it is the folder path we seek to create with
+    // extracted map contents.
     final Path extractionTarget =
-        ClientFileSystemHelper.getUserMapsFolder().resolve(extractionFolderName);
-
-    final boolean mapIsAlreadyExtracted = Files.exists(extractionTarget);
-    if (mapIsAlreadyExtracted) {
-      // no-op, we would not have expected for the map zip to have exist
-      return Optional.empty();
-    }
+        ClientFileSystemHelper.getUserMapsFolder()
+            .resolve(computeExtractionFolderName(mapZip.getFileName().toString()));
 
     log.info(
         "Extracting map zip: {} -> {}", mapZip.toAbsolutePath(), extractionTarget.toAbsolutePath());
@@ -126,22 +124,28 @@ public class ZippedMapsExtractor {
     final Path tempFolder = Files.createTempDirectory("triplea-unzip");
     ZipExtractor.unzipFile(mapZip, tempFolder);
 
-    // Check if the zip extracts to a single folder, if so, then to preserve pre-2.6 functionality
-    // we will use that as the map folder.
-    final Path folderToMove =
+    // Typically the next step is to move and rename the temp folder to the maps folder.
+    // But, if we just extracted exactly one folder, then we need to move and rename *that* folder.
+    final Path tempFolderWithExtractedMap =
         FileUtils.listFiles(tempFolder).size() == 1
+                && Files.isDirectory(FileUtils.listFiles(tempFolder).iterator().next())
             ? FileUtils.listFiles(tempFolder).iterator().next()
             : tempFolder;
 
-    // extraction done, now move the extracted folder to target location
-    Files.move(folderToMove, extractionTarget);
+    // replace extraction target folder contents with the temp folder containing the extracted zip
+    final boolean folderReplaced =
+        FileUtils.replaceFolder(tempFolderWithExtractedMap, extractionTarget);
+    if (!folderReplaced) {
+      return Optional.empty();
+    }
 
     // move properties file if it exists
     final Path propertiesFile =
         mapZip.resolveSibling(mapZip.getFileName().toString() + ".properties");
     if (Files.exists(propertiesFile)) {
       Files.move(
-          propertiesFile, extractionTarget.resolveSibling(extractionFolderName + ".properties"));
+          propertiesFile,
+          extractionTarget.resolveSibling(extractionTarget.getFileName() + ".properties"));
     }
 
     final boolean successfullyExtracted = Files.exists(extractionTarget);
@@ -158,7 +162,7 @@ public class ZippedMapsExtractor {
    * EG: 'map-name-master.zip' -> 'map-name'
    */
   @VisibleForTesting
-  static String createExtractionFolderName(final String mapZipName) {
+  static String computeExtractionFolderName(final String mapZipName) {
     String newName = mapZipName;
     if (newName.endsWith(".zip")) {
       newName = newName.substring(0, newName.length() - ".zip".length());
